+++ bench/scheduler_bench.ml
@@ -7,7 +7,8 @@ let () =
   Path.set_root (Path.External.cwd ());
   Path.set_build_dir (Path.Kind.of_string "_build")
 
-let prog = Option.value_exn (Bin.which "true")
+let prog =
+  Option.value_exn (Bin.which ~path:(Env.path Env.initial) "true")
 let run () = Process.run ~env:Env.initial Strict prog []
 
 let go ~jobs fiber =