+++ src/dune_env.mli
@@ -8,6 +8,7 @@ module Stanza : sig
     ; ocamlc_flags   : Ordered_set_lang.Unexpanded.t
     ; ocamlopt_flags : Ordered_set_lang.Unexpanded.t
     ; env_vars       : Env.t
+    ; binaries       : File_bindings.Unexpanded.t
     }
 
   type pattern =