+++ src/env_node.mli
@@ -18,3 +18,16 @@ val scope : t -> Scope.t
 val external_ : t -> profile:string -> default:Env.t -> Env.t
 
 val ocaml_flags : t -> profile:string -> expander:Expander.t -> Ocaml_flags.t
+
+val file_bindings
+  :  t
+  -> profile:string
+  -> expander:Expander.t
+  -> string File_bindings.t
+
+val artifacts
+  :  t
+  -> profile:string
+  -> default:Artifacts.t
+  -> expander:Expander.t
+  -> Artifacts.t