+++ src/menhir.ml
@@ -103,7 +103,8 @@ module Run (P : PARAMS) : sig end = struct
   (* Rule generation. *)
 
   let menhir_binary =
-    SC.resolve_program sctx "menhir" ~loc:None ~hint:"try: opam install menhir"
+    SC.resolve_program sctx ~dir "menhir" ~loc:None
+      ~hint:"try: opam install menhir"
 
   (* Reminder (from arg_spec.mli):